separated user and auth
