+++ src/controllers/user.js
@@ -1,0 +1,28 @@
+import { getUserById } from "../models/user.js";
+import { NotFoundError } from "../utils/exceptions.js";
+
+async function getUserController(req, res, next) {
+  try {
+    const { user_id } = req.params;
+    
+    if (!user_id) {
+      return next(new ValidationError("User invalid"));
+    }
+
+    const get_user = await getUserById(user_id);
+
+    if (!get_user) {
+      return next(new NotFoundError("User invalid"));
+    }
+
+    return res.status(200).json({
+      success: true,
+      message: "Found user with that id",
+      get_user,
+    });
+  } catch (error) {
+    return next(new ProgrammingError());
+  }
+}
+
+export { getUserController };
